Worker/requirements.txt
requests
pika
python-dotenv


